Added the signature of the spi_top level module.
The ports are still a work on progress, but as it currently stands, this should be everything needed for communication between the fs module and the spi module. Also added a .gitignore, which was missing for some unknown reason. Added just temporary files for the moment.

The spi module will be separated into 3 components: spi_top, spi_read, spi_write.
spi_top will be in charge of initialization and coordinating the two sub modules. The other two modules will be in charge of reading and writing.